+++ src/Makefile.am
@@ -173,6 +173,7 @@ BITCOIN_CORE_H = \
   evo/providertx.h \
   evo/simplifiedmns.h \
   evo/specialtx.h \
+  evo/specialtxman.h \
   dsnotificationinterface.h \
   governance/governance.h \
   governance/classes.h \
@@ -361,6 +362,7 @@ libdash_server_a_SOURCES = \
   evo/providertx.cpp \
   evo/simplifiedmns.cpp \
   evo/specialtx.cpp \
+  evo/specialtxman.cpp \
   flatfile.cpp \
   httprpc.cpp \
   httpserver.cpp \
