+++ src/rpc/rpcevo.cpp
@@ -11,6 +11,7 @@
 #include <evo/deterministicmns.h>
 #include <evo/simplifiedmns.h>
 #include <evo/specialtx.h>
+#include <evo/specialtxman.h>
 #include <index/txindex.h>
 #include <masternode/meta.h>
 #include <messagesigner.h>
@@ -291,7 +292,7 @@ static std::string SignAndSendSpecialTx(const CMutableTransaction& tx, bool fSub
     LOCK(cs_main);
 
     CValidationState state;
-    if (!CheckSpecialTx(CTransaction(tx), ::ChainActive().Tip(), state, ::ChainstateActive().CoinsTip())) {
+    if (!WITH_LOCK(cs_main, return CheckSpecialTx(CTransaction(tx), ::ChainActive().Tip(), state, ::ChainstateActive().CoinsTip()))) {
         throw std::runtime_error(FormatStateMessage(state));
     }
     } // cs_main
