+++ src/validation.cpp
@@ -47,6 +47,7 @@
 #include <masternode/payments.h>
 
 #include <evo/specialtx.h>
+#include <evo/specialtxman.h>
 #include <evo/deterministicmns.h>
 
 #include <llmq/instantsend.h>
