+++ test/lint/lint-circular-dependencies.sh
@@ -30,19 +30,13 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     # Dash
     "coinjoin/server -> net_processing -> coinjoin/server"
     "evo/cbtx -> evo/simplifiedmns -> evo/cbtx"
-    "evo/cbtx -> evo/specialtx -> evo/cbtx"
     "evo/deterministicmns -> evo/providertx -> evo/deterministicmns"
     "evo/deterministicmns -> evo/simplifiedmns -> evo/deterministicmns"
-    "evo/deterministicmns -> evo/specialtx -> evo/deterministicmns"
     "evo/deterministicmns -> llmq/commitment -> evo/deterministicmns"
     "evo/deterministicmns -> llmq/utils -> evo/deterministicmns"
     "evo/deterministicmns -> validation -> evo/deterministicmns"
     "evo/mnauth -> net_processing -> evo/mnauth"
-    "evo/specialtx -> llmq/blockprocessor -> evo/specialtx"
-    "evo/specialtx -> llmq/commitment -> evo/specialtx"
-    "evo/specialtx -> validation -> evo/specialtx"
     "governance/classes -> governance/governance -> governance/classes"
-    "governance/classes -> governance/governance -> net_processing -> masternode/payments -> governance/classes"
     "governance/governance -> governance/object -> governance/governance"
     "governance/governance -> masternode/sync -> governance/governance"
     "governance/governance -> net_processing -> governance/governance"
@@ -80,8 +74,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "qt/guiutil -> qt/optionsdialog -> qt/optionsmodel -> qt/guiutil"
     "evo/deterministicmns -> evo/simplifiedmns -> llmq/blockprocessor -> llmq/debug -> evo/deterministicmns"
     "evo/deterministicmns -> evo/simplifiedmns -> llmq/blockprocessor -> net_processing -> evo/deterministicmns"
-    "evo/specialtx -> llmq/blockprocessor -> net_processing -> txmempool -> evo/specialtx"
-    "evo/providertx -> evo/specialtx -> llmq/blockprocessor -> net_processing -> txmempool -> evo/providertx"
 
     "coinjoin/client -> net_processing -> coinjoin/client"
     "llmq/quorums -> net_processing -> llmq/quorums"
@@ -89,16 +81,13 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "llmq/dkgsession -> llmq/dkgsessionmgr -> llmq/dkgsession"
     "evo/deterministicmns -> validationinterface -> txmempool -> evo/deterministicmns"
     "llmq/chainlocks -> llmq/instantsend -> validation -> llmq/chainlocks"
-    "evo/deterministicmns -> evo/simplifiedmns -> llmq/blockprocessor -> net_processing -> masternode/payments -> evo/deterministicmns"
-    "governance/governance -> net_processing -> masternode/payments -> governance/governance"
     "coinjoin/coinjoin -> llmq/chainlocks -> net -> coinjoin/coinjoin"
     "evo/deterministicmns -> llmq/utils -> net -> evo/deterministicmns"
     "evo/deterministicmns -> llmq/utils -> net -> masternode/sync -> evo/deterministicmns"
-    "evo/mnhftx -> evo/specialtx -> evo/mnhftx"
     "policy/fees -> txmempool -> validation -> policy/fees"
     "policy/policy -> policy/settings -> policy/policy"
-    "bloom -> evo/specialtx -> llmq/blockprocessor -> net -> bloom"
-    "bloom -> evo/specialtx -> llmq/blockprocessor -> net_processing -> merkleblock -> bloom"
+    "evo/specialtxman -> validation -> evo/specialtxman"
+    "bloom -> llmq/commitment -> llmq/utils -> net -> bloom"
 )
 
 EXIT_CODE=0
